Rename numpy-loader -> npy-loader
