+++ modules/terrain/src/index.js
@@ -1,3 +1,3 @@
 export {TerrainLoader, TerrainWorkerLoader} from './terrain-loader';
 export {QuantizedMeshLoader, QuantizedMeshWorkerLoader} from './quantized-mesh-loader';
-export {NPYLoader as _NPYLoader, NPYWorkerLoader as _NPYWorkerLoader} from './numpy-loader';
+export {NPYLoader as _NPYLoader, NPYWorkerLoader as _NPYWorkerLoader} from './npy-loader';
