+++ modules/terrain/src/npy-loader.worker.js
@@ -1,4 +1,4 @@
-import {NPYLoader} from './numpy-loader';
+import {NPYLoader} from './npy-loader';
 import {createWorker} from '@loaders.gl/loader-utils';
 
 createWorker(NPYLoader);
